refactor: 연관된 사용자 정의 타입끼리 파일정리
- FruitStore와 JuiceMenu의 코드를 extension으로 정리
