--- JuiceMaker/JuiceMaker/Model/Fruit.swift
@@ -1,16 +1,0 @@
-//
-//  Fruit.swift
-//  JuiceMaker
-//
-//  Created by Jun Bang on 2021/10/20.
-//
-
-import Foundation
-
-enum Fruit: CaseIterable {
-    case strawberry
-    case bananna
-    case pineapple
-    case kiwi
-    case mango
-}
