--- JuiceMaker/JuiceMaker/Model/Ingredient.swift
@@ -1,13 +1,0 @@
-//
-//  Ingredient.swift
-//  JuiceMaker
-//
-//  Created by Jun Bang on 2021/10/20.
-//
-
-import Foundation
-
-struct Ingredient {
-    let fruit: Fruit
-    let amount: Int
-}
